Fix template error, add db sub-module
