+++ shakecast/app/db/db_data.py
@@ -1,0 +1,2 @@
+def main():
+    pass
